rename current-env
