+++ lib/current-env.js
