+++ lib/index.js
@@ -1,5 +1,5 @@
 const semver = require('semver')
-const currentEnv = require('./env')
+const currentEnv = require('./current-env')
 const { checkDevEngines } = require('./dev-engines')
 
 const checkEngine = (target, npmVer, nodeVer, force = false) => {
